\name{group_positive_r}
\alias{group_positive_r}
\title{
Function that sums the positive numbers within an array.
}
\description{
It takes an array as input and it spills out the array with the positive integers summed}
\arguments{
  \item{v}{an array of length  n>0}
}
\details{
The time complexity of this function is linear}
\value{
\item{res}{the new array}
}
\examples{
group_positive_r(c(5, -5, 3, -5, 10, 10, -8, 9, -5))

}
\keyword{v}