% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_algolia.R
\name{use_algolia}
\alias{use_algolia}
\title{Enable the Algolia Places JavaScript library}
\usage{
use_algolia()
}
\value{
called for the side effect of activating the Algolia Places JavaScript library
}
\description{
The function activates the capabilities of the Algolia Places JavaScript library
}
\examples{
\donttest{

# Set the algolia function at the beginning of the shiny app UI

use_algolia()



}
}
